renamed directories in pipeline_components to all follow "somethingizers" pattern
