+++ medacy/pipeline_components/__init__.py
@@ -1,10 +1,10 @@
 """Import pipeline components"""
 from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
-from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
+from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
 from medacy.pipeline_components.learners.bilstm_crf_learner import BiLstmCrfLearner
-from medacy.pipeline_components.tokenization.character_tokenizer import CharacterTokenizer
-from medacy.pipeline_components.tokenization.clinical_tokenizer import ClinicalTokenizer
-from medacy.pipeline_components.tokenization.systematic_review_tokenizer import SystematicReviewTokenizer
+from medacy.pipeline_components.tokenizers.character_tokenizer import CharacterTokenizer
+from medacy.pipeline_components.tokenizers.clinical_tokenizer import ClinicalTokenizer
+from medacy.pipeline_components.tokenizers.systematic_review_tokenizer import SystematicReviewTokenizer
 from medacy.pipeline_components.units.frequency_unit_component import FrequencyUnitComponent
 from medacy.pipeline_components.units.mass_unit_component import MassUnitComponent
 from medacy.pipeline_components.units.measurement_unit_component import MeasurementUnitComponent
