+++ medacy/pipelines/scispacy_pipeline.py
@@ -4,7 +4,7 @@ from spacy.tokenizer import Tokenizer
 
 from medacy.pipelines.base import BasePipeline
 from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
-from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
+from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
 from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
 from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
