+++ medacy/pipelines/systematic_review_pipeline.py
@@ -3,10 +3,10 @@ import spacy
 
 from medacy.pipelines.base.base_pipeline import BasePipeline
 from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
-from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
+from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
 from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
 from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
-from medacy.pipeline_components.tokenization.systematic_review_tokenizer import SystematicReviewTokenizer
+from medacy.pipeline_components.tokenizers.systematic_review_tokenizer import SystematicReviewTokenizer
 
 
 class SystematicReviewPipeline(BasePipeline):
